Add modules and update merge scripts to support dcos-commons import
Updates merge scripts for Cassandra, Elastic, HDFS, and Kafka (and their Beta equivalents). The existing scripts are updated to invoke the common 'dcos-commons-merge.sh', which handles the retrieval and conversion.

Adds the following modules to the processing chain to maintain compatibility with the Jekyll features that dcos-commons is using. All three are configured to only be enabled in the `services/` folder, and only take effect in content which explicitly invokes them. See `index.js`.

- `metalsmith-in-place-dcos`: This enables processing of `{{ mustache }}` statements within content. Rewrite of `metalsmith-in-place` which enables templating within files. This version uses a `render` annotation in the header, rather than guessing by filename extension.
- `metalsmith-include-content-dcos`: This enables including raw content from other files using C-like `#include path/to/file.txt` statements. Rewrite of `metalsmith-include-content` with the addition of support for specifying what files to process for include statements.
- `metalsmith-data-loader`: This allows fetching data models from external .json and .yml files, to be used when rendering the page.

Example
=======

Here's an example of how the above modules work together to allow templating of common content.

Input Files
-----------

- data.json:
  ```json
  {
    "type": "shared content",
    "name": "world"
  }
  ```

- template.tmpl:
  ```
  Hello, {{ model.name }}!
  ```

- index.md:
  ```
  ---
  model: path/to/data.json
  render: mustache
  ---
  Now for some {{ model.type }}:
  #include path/to/template.tmpl
  ```

Result
------

```
Now for some shared content:
Hello, world!
```